+++ hisho_core/src/lib.rs
@@ -3,8 +3,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-pub mod build;
-pub mod config;
+pub mod build_tool;
+pub mod environment;
 pub mod config_models;
 pub mod containers;
 pub mod files;