+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Core: Renamed `config` module to `environment` module
 * Core: Renamed `build` module to `build_tool` module
 * Core: Make `resolve_files_from_globs` in `build_tool` module public
+* Core: Make `get_home_dir` in `files` module public
 
 <!--
 ### Added